`feat` change to Long

+++ features/content/play/src/main/java/com/tokopedia/play/data/Product.kt
@@ -23,7 +23,7 @@ data class Product(
         @SerializedName("price_formatted")
         val priceFormatted: String = "",
         @SerializedName("discount")
-        val discount: Int = 0,
+        val discount: Long = 0,
         @SerializedName("order")
         val order: Int = 0,
         @SerializedName("is_variant")
