+++ features/content/play/src/main/java/com/tokopedia/play/view/custom/multiplelikes/PlayLikeBubblesManager.kt
@@ -51,8 +51,8 @@ class PlayLikeBubblesManager(
     }
 
     fun shot(
-        likeAmount: Int,
-        shotPerBatch: Int,
+        likeAmount: Long,
+        shotPerBatch: Long,
         prioritize: Boolean = false,
         delayPerBatchInMs: Long = 0L,
         reduceOpacity: Boolean = false,
@@ -62,7 +62,7 @@ class PlayLikeBubblesManager(
             for (i in 1..likeAmount) {
                 if (delayPerBatchInMs > 0) delay(delayPerBatchInMs)
                 for(j in 1..shotPerBatch) {
-                    if (j != 1) delay(DEFAULT_DELAY)
+                    if (j != 1L) delay(DEFAULT_DELAY)
                     val chosenBubble = bubbleList.random()
                     shotInternal(
                         chosenBubble.icon,
