+++ features/content/play/src/main/java/com/tokopedia/play/view/viewmodel/PlayViewModel.kt
@@ -495,7 +495,7 @@ class PlayViewModel @AssistedInject constructor(
             if (videoLatencyPerformanceMonitoring.hasStarted) {
                 videoLatencyPerformanceMonitoring.stop()
                 val durationInSecond = videoLatencyPerformanceMonitoring.totalDuration / DURATION_DIVIDER
-                playLog.logTimeToFirstByte(durationInSecond.toInt())
+                playLog.logTimeToFirstByte(durationInSecond.toLong())
             }
         }
 
@@ -1543,7 +1543,7 @@ class PlayViewModel @AssistedInject constructor(
                 } else if (diffLike > 0) {
                     _uiEvent.emit(
                         ShowLikeBubbleEvent.Single(
-                            diffLike.toInt(),
+                            diffLike,
                             reduceOpacity = true,
                             config = _likeInfo.value.likeBubbleConfig,
                         )
@@ -2522,7 +2522,7 @@ class PlayViewModel @AssistedInject constructor(
         private const val ONBOARDING_DELAY = 5000L
         private const val INTERACTIVE_FINISH_MESSAGE_DELAY = 2000L
 
-        private const val LIKE_BURST_THRESHOLD = 30
+        private const val LIKE_BURST_THRESHOLD = 30L
 
         /**
          * Request Code When need login
